Updated "virtual LocationVector GetStartingCoords(uint8 Team)" function to "virtual LocationVector GetStartingCoords(uint8 Team) = 0"
Warsong gulch code tweaks
* Moved buffs, flags, start positions to header
* Removed SpawnBuff function
* Implemented SpawnBgGameObject (to make game object spawn stuff a lot easer)
* Replaced many magic numbers with definitions with some describtions

Some various warnings fixes

+++ src/arcemu-world/Battleground.h
@@ -135,7 +135,7 @@ class SERVER_DECL CBattleground : public EventableObject
         void SendPVPData(Player* plr);
 
         /* Get the starting position for this team. */
-        virtual LocationVector GetStartingCoords(uint32 Team) = 0;
+        virtual LocationVector GetStartingCoords(uint8 Team) = 0;
 
         /* Send a packet to the entire battleground */
         void DistributePacketToAll(WorldPacket* packet);
